build(dist): removal of dist:js script
This commit re-removes the dist:js script that was added
back in the previous commit. This script is not needed
as the index.js file points to the included ./src dir.

The npm ignore is also set to only ignore Scss and CSS
resource files. The processed css-js files are included.

Changes to be committed:
modified:   package.json