fix: update small CTA to adjust content alignment #191
There was mainly an issue with the placement of an icon within the small
CTA use case. The update allows for easier management of the additional
content.

Stylelint rules were adjusted to allow for new code without
having to also do a code refactor to meet linting needs.

Changes to be committed:
modified:   src/style.scss