perf: update legacy .is-touch to media query
This commit updates all the legacy .is-touch selector references, that
was deprecated a while ago, to use the preferred @media hover query.

@include auro_anchorButton was also updated to use the `noncomponent`
API option. While this is inconsistent with intended use, the
`component` value of the API calls in the legacy .is-touch reference
that is no longer supported.

This is discussed in the following WCSS issue
https://github.com/AlaskaAirlines/WebCoreStyleSheets/issues/192

Changes to be committed:
modified:   src/component-base.mjs
modified:   src/style.scss