fix: adjust shape of UI for a11y UI 209
The issue around the misshaped UI was related to the manual nudging of
the inserted icon and the updates to the icons shape and relative size
over time.

Updating the nudge from 2px to 4px addresses the UI issue.

This commit also has an update to the README based on updates to the
generator with the build process.

On branch dsande/focusVisible/209
Changes to be committed:
modified:   README.md
modified:   src/style.scss