feat: default dynamic HREF support #243
The default behavior now dynamically adjusts a URL's `hostname` to
match the current page if the `origin` attribute is not specified,
as the new default for `origin` is `dynamic`. Previously, URLs
relied solely on the `hostname` defined in the `href` attribute,
but with this change, links will adapt to the environment
(e.g., development, staging, or production).

The purpose of this update was to allow for dynamic resetting of
absolute HREFs when links are intended to work within the new
multi-theme experiences.

On branch dsande/dynamichostname/243
Changes to be committed:
modified:   src/auro-hyperlink.js
modified:   src/component-base.mjs